+++ CHANGELOG.md
@@ -14,7 +14,8 @@ This is the first release that supports only MPD 0.23.5 and higher and Lua 5.4.x
 
 - Feat: Use myGPIOd REST-API #1510
 - Feat: Implement merge sort for linked lists
-- Feat: Bump requirement for MPD and Lua versions
+- Feat: Regularly save the myMPD state if myMPD is active
+- Upd: Bump requirement for MPD and Lua versions
 
 ***